chore: rename a bunch of files
