+++ pyannote/audio/pipelines/__init__.py
@@ -23,8 +23,8 @@
 from .multilabel import MultiLabelSegmentation
 from .overlapped_speech_detection import OverlappedSpeechDetection
 from .resegmentation import Resegmentation
-from .segmentation import SpeakerSegmentation
 from .speaker_diarization import SpeakerDiarization
+from .speaker_segmentation import SpeakerSegmentation
 from .voice_activity_detection import VoiceActivityDetection
 
 __all__ = [
