+++ pyannote/audio/tasks/__init__.py
@@ -21,7 +21,7 @@
 # SOFTWARE.
 
 from .segmentation.multilabel import MultiLabelSegmentation  # isort:skip
-from .segmentation.diarization import SpeakerDiarization  # isort:skip
+from .segmentation.speaker_diarization import SpeakerDiarization  # isort:skip
 from .segmentation.voice_activity_detection import VoiceActivityDetection  # isort:skip
 from .segmentation.overlapped_speech_detection import (  # isort:skip
     OverlappedSpeechDetection,
